+++ frontend/app/app.vue
@@ -1,5 +1,11 @@
+<script setup lang="ts">
+// const { isLoading } = useLoadingIndicator()
+</script>
+
 <template>
 	<UApp>
+		<NuxtLoadingIndicator />
+
 		<NuxtLayout>
 			<NuxtPage />
 		</NuxtLayout>